Moved social methods to a separate social-utils.js
